src/main/java/com/github/brunomndantas/jscrapper/support/parser/collection/text/reference/CollectionBooleanTextParser.java
package com.github.brunomndantas.jscrapper.support.parser.collection.text.reference;

import com.github.brunomndantas.jscrapper.support.parser.collection.text.CollectionTextParser;
import org.openqa.selenium.WebElement;

public class CollectionBooleanTextParser extends CollectionTextParser {

    public CollectionBooleanTextParser() {
        super(null);
    }



    @Override
    protected Object parseElement(WebElement element) throws Exception {
        return Boolean.parseBoolean(element.getText());
    }

}
